wap-backend-azure/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
sentence-transformers==2.7.0
qdrant-client==1.7.0
python-dotenv==1.0.0
redis==5.0.1

# Azure SDKs
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-identity==1.15.0
msal==1.26.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Linting and formatting
ruff==0.1.6
black==23.11.0
mypy==1.7.1
